Code/streamlit-movie-app/requirements.txt
streamlit==1.38.0
pyspark==3.5.1
azure-storage-blob